refactor(neuron-ui): move localCache into services
